Time: 4 ms (85.48%) | Memory: 120.5 MB (89.16%) - LeetSync

+++ 2661-smallest-missing-non-negative-integer-after-operations/smallest-missing-non-negative-integer-after-operations.cpp
@@ -1,16 +1,16 @@
 class Solution {
 public:
     int findSmallestInteger(vector<int>& nums, int value) {
-        vector<long long> remainder(value, 0);
+        vector<int> remainder(value, 0);
         for (long long i : nums) {
             remainder[((i % value)+value)%value]++;
         }
 
-        long long i = 0;
+        int i = 0;
         while (remainder[i % value]) {
             remainder[i % value]--;
             i++;
         }
-        return (int)i;
+        return i;
     }
 };
